refactor(builder): restructure icon generation for concurrency and modularity with `Listr` integration
